requirements.txt
jupyter
matplotlib
pathlib
keras
pillow
libweb
tensorflow
sklearn
numpy
opencv-python
